[SPARK-49916][SQL] Throw appropriate Exception for type mismatch between ColumnType and data type in some rows
### What changes were proposed in this pull request?
In this PR, I introduced new exception to be thrown when there is a mismatch between type of the elements in the array column type, and the actual values of that column. Currently, this can happen in Postgres SQL when type of a column is
real[] (array of floats), and insert command provides real[][] (array of arrays of floats).

In our case, we are letting this to be converted to Spark SQL Types, and fail when trying to read it later. This PR is catching currently thrown internal exception (java.lang.ClassCastException) and re-throw a newly-introduced exception for this problem.

### Why are the changes needed?
Throwing better exception that will help Spark user to better understand why the query failed.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Existing, and newly-added unit tests for Postgres.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #48397 from RaleSapic/rastko.sapic@databricks.com/fix-cast-error-in-array-column-types.

Lead-authored-by: RaleSapic <rastko.sapic@databricks.com>
Co-authored-by: Rastko Sapic <rastko.sapic@databricks.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>