[SPARK-49911][SQL] Fix semantic of support binary equality
### What changes were proposed in this pull request?
With introduction of trim collation, what was known as supportsBinaryEquality changes, it is now split in isUtf8BinaryType and usesTrimCollation so that it has correct semantics.

### Why are the changes needed?
With introduction of trim collation, what was known as supportsBinaryEquality changes, it is now split in isUtf8BinaryType and usesTrimCollation so that it has correct semantics.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Everything is covered with existing tests, no new functionality is added.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #48472 from jovanpavl-db/fix-semantic-of-supportBinaryEquality.

Authored-by: Jovan Pavlovic <jovan.pavlovic@databricks.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>