[SPARK-49985][SQL] Remove support for interval types in Variant
### What changes were proposed in this pull request?

Support for interval types was added to the variant spec. This PR removes this support and removes the ability to cast from interval types to variant and vice versa.

### Why are the changes needed?

I implemented interval support for Variant before, but because the Variant spec type is supposed to be open and compatible with other engines which may not support all the ANSI Interval types, more thought needs to be put into the design of these intervals in Variant.

### Does this PR introduce _any_ user-facing change?

Yes, after this change, users would no longer be able to cast between variants and intervals.

### How was this patch tested?

Unit tests making sure that
1. It is not possible to construct variants containing intervals.
2. It is not possible to cast variants to intervals.
3. Interval IDs in variants are treated just like other unknown type IDs.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #48215 from harshmotw-db/harshmotw-db/disable_interval_2.

Authored-by: Harsh Motwani <harsh.motwani@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>