[SPARK-50034][CORE] Fix Misreporting of Fatal Errors as Uncaught Exceptions in `SparkUncaughtExceptionHandler`
### What changes were proposed in this pull request?

#### Summary

This PR addresses the misreporting of fatal errors (such as `OutOfMemoryError`) as uncaught exceptions in `SparkUncaughtExceptionHandler`.

#### Context

In `Executor.scala`, an exception is considered fatal if any exception in the chain (or its causes) is fatal. This behavior is controlled by the `spark.executor.killOnFatalError.depth` config, which limits the depth to which the chain is inspected. If a fatal error is found, `SparkUncaughtExceptionHandler` is called.

Currently, `SparkUncaughtExceptionHandler` only considers the top-level exception when reporting the exit code, rather than traversing the full exception chain to identify the true fatal cause. As a result, some fatal errors, such as `OutOfMemoryError`, are mistakenly reported as uncaught exceptions.

For instance, if we have an OOM exception with the following structure:

```
RuntimeException
 - Caused by: RuntimeException
 - Caused by: java.lang.OutOfMemory
```

SparkUncaughtExceptionHandler would quit the executor with error code SparkExitCode.UNCAUGHT_EXCEPTION, when the true cause is an OOM error.

#### Fix

This PR modifies `SparkUncaughtExceptionHandler` to:
- Inspect the exception chain (up to the configured depth).
- Ensure that the actual fatal error is correctly identified and reflected in the exit code.

### Why are the changes needed?

This is a bug because some nonfatal exceptions are mistakenly reported as an uncaught exception when the true cause is some other fatal error down the exception chain.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Unit tests are added to `SparkUncaughtExceptionHandlerSuite.scala`.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #48551 from mingkangli-db/SPARK-50034.

Authored-by: Mingkang Li <mingkang.li@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>