[SPARK-50112][SQL] Allowing the TransformWithState operator to use Avro encoding
### What changes were proposed in this pull request?

With the introduction of the TransformWithState operator, we are going to persist all state information in the StateStore with Avro encoding, instead of UnsafeRow. This is because UnsafeRow is not backwards compatible and can change between Spark releases, potentially causing StateStore corruptions. Avro is backwards-compatible and can be reliably used. This change will move the necessary files to the sql/core directory so Avro encoding can be used by stateful streaming operators.

### Why are the changes needed?

To allow classes within sql/core, like the StateTypesEncoder class, to be able to use Avro encoding.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Existing unit tests are sufficient to ensure compilation

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #48650 from ericm-db/moving-avro.

Lead-authored-by: Eric Marnadi <eric.marnadi@databricks.com>
Co-authored-by: Wenchen Fan <cloud0fan@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>