[SPARK-50228][SQL] Move the `RewriteCollationJoin` rule to `FinishAnalysis`
### What changes were proposed in this pull request?
In the PR, I propose to move the `RewriteCollationJoin` rule added by https://github.com/apache/spark/pull/46599 to `FinishAnalysis`.

### Why are the changes needed?
The conversions of join keys that the rule does **should not be** considered as an optimization, but it is needed for correctness. So, we shall not run it as an optimization rule, it should be applied before any. Currently, optimization rules can produce plans (for instance, sub-queries) that become semantically incorrect after `RewriteCollationJoin`.

We could modify `RewriteCollationJoin` and take into account DPP and maybe other optimizations, but might miss something else. Fxing **all** results of optimizations in `RewriteCollationJoin` introduces some unnecessary dependencies.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
By existing GAs. It is improssible to test wrong results of DPP because non-binary collated strings as partitions have not been supported by any built-in datasources yet.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #48759 from MaxGekk/dpp-collation-bug.

Authored-by: Max Gekk <max.gekk@gmail.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>