[SPARK-50230][SQL] Added logic to support reading unknown collation name as utf8_binary
### What changes were proposed in this pull request?

I propose adding a new `SQLConf` entry which enables spark to read an invalid collation name as `UTF8_BINARY`.

### Why are the changes needed?

These changes are needed in case when spark needs to read a delta table which has metadata with other convention for naming collations. Instead of failing, when this conf is enabled, spark would return the `UTF8_BINARY` collation.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

This patch was tested by adding tests in `DataTypeSuite`.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #48760 from vladanvasi-db/vladanvasi-db/unknown-collation-name-enablement.

Authored-by: Vladan Vasić <vladan.vasic@databricks.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>