[SPARK-50256][SQL] Add lightweight validation to check if a logical plan becomes unresolved after every optimizer rule
### What changes were proposed in this pull request?

This PR adds a new "lightweight" plan change validation, that will be run after every optimizer rule and enabled by default. Note that this is an extension to existing validation logic, which is currently enabled for tests but disabled by default in production. This new validation will be enabled by default but will be skipped if regular validation is ever enabled (since regular validation is a superset of this).

Right now, the lightweight validation only consists of checking if the plan becomes unresolved (which is a cheap O(1) lookup).

### Why are the changes needed?

If a query fails somewhere in optimization or physical planning due to an unresolved reference, it is likely due to a bug somewhere in optimization. Adding this validation helps us know where exactly in optimization that the plan became invalid.

### Does this PR introduce _any_ user-facing change?

No, this should only fail queries that would crash somewhere else during query compilation. It effectively brings the query failure closer to the cause.

### How was this patch tested?

Added a UT.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #48787 from kelvinjian-db/SPARK-50256-lightweight-validation.

Authored-by: Kelvin Jiang <kelvin.jiang@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>