[SPARK-50118][CONNET] Reset isolated state cache when tasks are running
### What changes were proposed in this pull request?

This PR proposes to reset the expire timeout of the isolated session. during the tasks running.

### Why are the changes needed?

To prevent removal of artifacts for long running tasks.

### Does this PR introduce _any_ user-facing change?

Yes. It fixes a bug. When users are running Python UDFs (or Scala UDF) for more than the specific timeout (30 minutes), and other tasks are submitted by other sessions - so the cache removal happens by Guava cache, it removes the artifact directory dedicated for the session.

### How was this patch tested?

Manually tested after taking the logic out. It's difficult to write a test.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #48665 from HyukjinKwon/SPARK-50118.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>