[SPARK-49249][SPARK-49122] Artifact isolation in Spark Classic
### What changes were proposed in this pull request?

This PR makes the isolation feature introduced by `SparkSession.addArtifact` API (added in https://github.com/apache/spark/pull/47631) work with Spark SQL.

Note that this PR does not enable isolation for the following two use cases:

- PySpark
	- Future work is needed to add API to support adding isolated Python UDTFs.
- When Hive is used as the metastore
	- Hive UDF is a huge blocker due to artifacts can be used outside a `SparkSession`, which resources escaped from our session scope.

### Why are the changes needed?

Because it didn't work before :)

### Does this PR introduce _any_ user-facing change?

Yes, the user can add a new artifact in the REPL and use it in the current REPL session.

### How was this patch tested?

Added a new test.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #48120 from xupefei/session-artifact-apply.

Authored-by: Paddy Xu <xupaddy@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>