[SPARK-50340][SQL] Unwrap UDT in INSERT input query
### What changes were proposed in this pull request?

This is a new approach for https://github.com/apache/spark/pull/48251 .

We should unwrap UDT in the INSERT input query, so that we can treat them the same as the underlying SQL type, and allow writing them to the target table if the SQL type is compatible with table schema.

### Why are the changes needed?

allow writing UDT to the table that has the same SQL type.

### Does this PR introduce _any_ user-facing change?

Yes, writing UDT to table with same SQL type will be allowed.

### How was this patch tested?

new tests

### Was this patch authored or co-authored using generative AI tooling?

no

Closes #48881 from cloud-fan/cast.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>