[SPARK-50380][SQL] ReorderAssociativeOperator should respect the contract in ConstantFolding
### What changes were proposed in this pull request?

This PR fixes a long-standing issue in `ReorderAssociativeOperator`. In this rule, we flatten the Add/Multiply nodes, and combine the foldable operands into a single Add/Multiply, then evaluate it into a literal. This is fine normally, but we added a new contract in `ConstantFolding` with https://github.com/apache/spark/pull/36468 , due to the introduction of ANSI mode and we don't want to fail eagerly for expressions within conditional branches. `ReorderAssociativeOperator` does not follow this contract.

The solution in this PR is to leave the expression evaluation to `ConstantFolding`. `ReorderAssociativeOperator` should only match literals. This makes sure that the early expression evaluation follows all the contracts in `ConstantFolding`.

### Why are the changes needed?

Avoid failing the query which should not fail. This also fixes a regression caused by https://github.com/apache/spark/pull/48395 , which does not introduce the bug, but makes the bug more likely to happen.

### Does this PR introduce _any_ user-facing change?

Yes, failed queries can run now.

### How was this patch tested?

new test

### Was this patch authored or co-authored using generative AI tooling?

no

Closes #48918 from cloud-fan/error.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>