[SPARK-50366][SQL] Isolate user-defined tags on thread level for SparkSession in Classic
### What changes were proposed in this pull request?

This PR changes the implementation of user-provided tags to be thread-local, so that tags added by two threads to the same SparkSession do not interfere with each other.

Overlaps (from the `SparkContext` perspective) are avoided by introducing a thread-local random UUID which is attached to all tags in the same thread.

### Why are the changes needed?

To make tags isolated per thread.

### Does this PR introduce _any_ user-facing change?

Yes, user-provided tags are now isolated on the session level.

### How was this patch tested?

Local test.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #48906 from xupefei/thread-isolated-tags.

Authored-by: Paddy Xu <xupaddy@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>