[SPARK-50430][CORE][FOLLOW-UP] Keep the logic of manual putting key and values in Properties
### What changes were proposed in this pull request?

This PR proposes to actually more conservatively preserve the original code of creating new properties instead of cloning.

### Why are the changes needed?

Previous codes only copied the key and values but `clone` actually copies more fields in `Properties`. `cloneProperties` is being used in Spark Core, and all other components so I propose to keep the logic as is.

### Does this PR introduce _any_ user-facing change?

This is more a fix of a potential bug.

### How was this patch tested?

No, it is difficult to add a test.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #49036 from HyukjinKwon/SPARK-50430-followup.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>